+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   any error while disconnecting as a disconnect error.
 * Remove the need to pass `onConnectionError` handler to the `Transport.connect(...)` method and instead treat any
   error while connecting as a connect error.
+* Remove the need to pass `onDisconnection` handler to the `Transport.disconnect(...)` method and directly invoke
+  callback from `Transport`.
 
 ### [v6.37](https://github.com/replicant4j/replicant/tree/v6.37) (2019-03-18)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.36...v6.37)